+++ gdtoa/g_Qfmt_p.c
@@ -70,7 +70,7 @@ g_Qfmt_p(char *buf, void *V, int ndig, size_t bufsize, int nik)
 
 	if (ndig < 0)
 		ndig = 0;
-	if (bufsize < ndig + 10)
+	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
 	L = (ULong*)V;
