+++ gdtoa/g_ddfmt.c
@@ -60,7 +60,7 @@ g_ddfmt(char *buf, double *dd0, int ndig, size_t bufsize)
 #define Rounding FPI_Round_near
 #endif /*}}*/
 
-	if (bufsize < 10 || bufsize < ndig + 8)
+	if (bufsize < 10 || bufsize < (size_t)(ndig + 8))
 		return 0;
 
 	dd = (U*)dd0;
@@ -152,7 +152,7 @@ g_ddfmt(char *buf, double *dd0, int ndig, size_t bufsize)
 		}
 	mode = 2;
 	if (ndig <= 0) {
-		if (bufsize < (int)(fpi.nbits * .301029995664) + 10) {
+		if (bufsize < (size_t)((int)(fpi.nbits * .301029995664) + 10)) {
 			Bfree(z);
 			return 0;
 			}
